some cleanup, starting shadow investigation

+++ threepp/quick/objects/Box.h
@@ -17,11 +17,11 @@ namespace quick {
 class Box : public ThreeQObject
 {
   Q_OBJECT
-  Q_PROPERTY(unsigned width READ width WRITE setWidth NOTIFY widthChanged)
-  Q_PROPERTY(unsigned height READ height WRITE setHeight NOTIFY heightChanged)
-  Q_PROPERTY(unsigned depth READ depth WRITE setDepth NOTIFY depthChanged)
+  Q_PROPERTY(float width READ width WRITE setWidth NOTIFY widthChanged)
+  Q_PROPERTY(float height READ height WRITE setHeight NOTIFY heightChanged)
+  Q_PROPERTY(float depth READ depth WRITE setDepth NOTIFY depthChanged)
 
-  unsigned _width=0, _height=0, _depth=0;
+  float _width=0, _height=0, _depth=0;
 
   DynamicMesh::Ptr _mesh;
 
@@ -41,11 +41,11 @@ protected:
 public:
   Box(QObject *parent = nullptr) : ThreeQObject(parent) {}
 
-  unsigned width() const {return _width;}
-  unsigned height() const {return _height;}
-  unsigned depth() const {return _depth;}
+  float width() const {return _width;}
+  float height() const {return _height;}
+  float depth() const {return _depth;}
 
-  void setWidth(unsigned width) {
+  void setWidth(float width) {
     if(_width != width) {
       _width = width;
       if(_mesh) _mesh->geometry_t<geometry::Box>()->setWidth(width);
@@ -53,7 +53,7 @@ public:
     }
   }
 
-  void setHeight(unsigned height) {
+  void setHeight(float height) {
     if(_height != height) {
       _height = height;
       if(_mesh) _mesh->geometry_t<geometry::Box>()->setHeight(height);
@@ -61,7 +61,7 @@ public:
     }
   }
 
-  void setDepth(unsigned depth) {
+  void setDepth(float depth) {
     if(_depth != depth) {
       _depth = depth;
       if(_mesh) _mesh->geometry_t<geometry::Box>()->setDepth(depth);
